fix: double quantity and this on private fields

+++ src/it/unibo/oop/lab05/ex3/ProductImpl.java
@@ -3,9 +3,9 @@ package it.unibo.oop.lab05.ex3;
 public class ProductImpl implements Product {
 
 	private final String name;
-	private int quantity;
+	private double quantity;
 
-	public ProductImpl(String name, int quantity) {
+	public ProductImpl(String name, double quantity) {
 		this.name = name;
 		this.quantity = quantity;
 	}
